Remove runEvolution and getGenerationsIterator
